refactor: remove each specific config's folder
