+++ src/config/app-config.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
-import appConfig from './app/app.config';
-import authConfig from './auth/auth.config';
-import databaseConfig from './database/database.config';
-import frontendConfig from './frontend/frontend.config';
+import appConfig from './app.config';
+import authConfig from './auth.config';
+import databaseConfig from './database.config';
+import frontendConfig from './frontend.config';
 
 @Module({
   imports: [
