+++ src/config/typeorm/typeorm-config.service.ts
@@ -2,8 +2,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import { ConfigType } from '@nestjs/config';
 import { TypeOrmModuleOptions, TypeOrmOptionsFactory } from '@nestjs/typeorm';
 
-import databaseConfig from '../database/database.config';
 import { entitiesPath, migrationsPath } from './typeorm-config';
+import databaseConfig from '../database.config';
 
 @Injectable()
 export class TypeOrmConfigService implements TypeOrmOptionsFactory {
